+++ packages/client/src/pages/Invoice/InvoiceDetails.tsx
@@ -11,11 +11,11 @@ import { FormDialog } from '@components/form/FormDialog';
 import { useState } from 'react';
 import { GetAllInvoicesDocument, SearchInvoiceDocument, useCreateOrUpdateInvoiceMutation, useSearchInvoiceQuery } from '@graphql/invoice/invoice';
 import { DisplayCard } from '@components/DisplayCard.component';
-import { CommentInputBox } from '@pages/Invoice/CommentInputBox';
+import { CommentInputBox } from '@pages/Invoice/components/comment/CommentInputBox';
 import { useAddCommentMutation, useDeleteCommentMutation } from '@graphql/comment/comment';
-import { CommentDisplayComponent } from '@pages/Invoice/CommentDisplayComponent';
-import { CommentListItem } from '@pages/Invoice/CommentListItem';
-import { CommentList } from '@pages/Invoice/CommentList';
+import { CommentDisplayComponent } from '@pages/Invoice/components/comment/CommentDisplayComponent';
+import { CommentListItem } from '@pages/Invoice/components/comment/CommentListItem';
+import { CommentList } from '@pages/Invoice/components/comment/CommentList';
 import * as Yup from 'yup';
 import { Form, Formik } from 'formik';
 import { ObserverTextInput } from '@components/form/ObserverTextInput';
